Rename `db_cleanup` to `deletion_planner`, flesh out docstrings, and lint.
