--- robot-server/tests/test_db_cleanup.py
@@ -1,3 +1,0 @@
-def test_db_cleanup() -> None:
-    # TODO: Add tests for db_cleanup.py.
-    raise NotImplementedError()
