+++ robot-server/tests/test_deletion_planner.py
@@ -1,0 +1,6 @@
+"""Unit tests for `deletion_planner`."""
+
+
+def test_db_cleanup() -> None:  # noqa: D103
+    # TODO: Add tests for db_cleanup.py.
+    raise NotImplementedError()
